requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlmodel==0.0.16
aiosqlite==0.20.0
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3
redis==5.0.1

# Templates & Forms
jinja2==3.1.3
python-multipart==0.0.9

# HTTP & Crawling
httpx==0.27.0
beautifulsoup4==4.12.3
trafilatura==1.6.0

# AI
openai==1.12.0

# Configuration
python-dotenv==1.0.1
pydantic-settings==2.2.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.1.0.post1
authlib==1.3.0
itsdangerous==2.1.2

# Payment - Stripe
stripe==8.4.0

# Background Tasks & Queues
celery==5.3.6

# Monitoring & Logging
sentry-sdk[fastapi]==1.40.4

# Server
gunicorn==21.2.0

# Testing (dev only)
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0
